+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Next Release
 ------------
 * Fix: Better error on bad ModelItem constructor argument (#50)
 * Fix: Suppress archiving of downloaded workspaces by setting archive location to None (#54)
+* Fix: Add DeploymentView (#55)
 
 0.2.1 (2020-11-27)
 ------------------